Rename error code files
