+++ aurora_data_api/error_codes_mysql.py
